PIQA-64 - Purpose update draft mode receive (#80)
